#1432 Removed centre managers' isMetaValuePresent / markMetaValuePresent markers (and their impl) used previously when centre managers have been serialised and stored in database. Moved MetaValueType to better place where it is used now (CentreUpdater infrastructure). Removed last usage of deep copy method. Removed method of converting old centre-manager-based diff object to new version bag-like diff object.
