+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/entity_centre/review/criteria/EntityQueryCriteria.java
@@ -115,16 +115,6 @@ public abstract class EntityQueryCriteria<C extends ICentreDomainTreeManagerAndE
     }
 
     /**
-     * Returns the copy of centre domain tree manager
-     *
-     * @return
-     */
-    //TODO remove this later after details will be implemented using
-    public C getCentreDomainTreeManagerAndEnhnacerCopy() {
-        return EntityUtils.deepCopy(cdtme, serialiser);
-    }
-
-    /**
      * Returns the root for which this criteria was generated.
      *
      * @return
